extract credit payment
