+++ environment/__init__.py
@@ -1,4 +1,5 @@
 from .environment import ProductOwnerEnv
+from .credit_payer_env import CreditPayerEnv
 from environment.StochasticGameStartEnv import StochasticGameStartEnv
